aizen-skill-bridge/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "aizen-skill-bridge"
version = "0.1.0"
description = "Python skill runtime bridge for Aizen Agent"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
aizen-skill = "aizen_skill_bridge.cli:main"

[tool.setuptools.packages.find]
include = ["aizen_skill_bridge*"]